Fixed all Typings & Updated NotCertifiedKibana to include all methods inside the class
