+++ compose.yml
@@ -7,6 +7,8 @@ services:
       - 5678:5678
     env_file:
       - .env
+    environment:
+      - PYTHONPATH=/app/cat/plugins/kibcat/imports
     volumes:
       # Cheshire Cat
       - ./cat/static:/app/cat/static
