+++ mypy.ini
@@ -1,6 +1,14 @@
 [mypy]
+mypy_path = src
 python_version = 3.13
-ignore_missing_imports = True
-strict = True
+ignore_missing_imports = true
+strict = true
+disallow_untyped_defs = false
+explicit_package_bases = true
+namespace_packages = true
+no_site_packages = true
 exclude = cat/
-disallow_untyped_defs = False
+
+[mypy-prison]
+ignore_missing_imports = true
+
