+++ pyproject.toml
@@ -1,7 +1,18 @@
 [tool.black]
-line-length = 100
+line-length = 120
 target-version = ['py310']
 
 [tool.isort]
 profile = "black"
-line_length = 100
+line_length = 120
+
+#[mypy]
+#python_version = "3.13"
+#ignore_missing_imports = true
+#strict = true
+#disallow_untyped_defs = false
+#explicit_package_bases = true
+#exclude = "cat/"
+
+#[tool.mypy.prison]
+#ignore_missing_imports = true
