+++ pytest.ini
@@ -1,5 +1,7 @@
 [pytest]
-minversion = 6.0
-addopts = -ra
-testpaths =
-	tests
+pythonpath = src
+minversion = "6.0"
+addopts = "-ra"
+testpaths = [
+    "tests",
+]
