+++ src/kiblog/__init__.py
@@ -1,0 +1,3 @@
+from .base_logger import BaseLogger
+
+__all__ = ["BaseLogger"]
