+++ src/kibtypes/__init__.py
@@ -1,0 +1,3 @@
+from .parsed_kibana_url import ParsedKibanaURL
+
+__all__ = ["ParsedKibanaURL"]
