+++ src/kibtypes/parsed_kibana_url.py
